chore(clippy): make the tree clean under -D warnings
Mostly `cargo clippy --fix`: is_some_and, div_ceil, the Default impls
next to an argument-less new(), module docs where a doc comment sat
above a use, resize() where a loop pushed the same item.

Indexing parallel arrays by position is deliberate in the codecs, so
needless_range_loop is allowed crate-wide with that reason; the two
benchmark row tuples and the container parse keep their shape behind
type_complexity allows.

Also drops an unused compress_global helper in tests/v10_standard_bench.rs
and ignores lab/.

+++ src/codecs/lzma_style.rs
@@ -35,7 +35,7 @@ fn dist_slot(dist: u32) -> u32 {
         return dist;
     }
     let bsr = 31 - dist.leading_zeros(); // floor(log2(dist))
-    ((bsr as u32) << 1) + ((dist >> (bsr - 1)) & 1)
+    (bsr << 1) + ((dist >> (bsr - 1)) & 1)
 }
 
 /// Convert a distance slot back to the base distance and number of extra bits.
